This is the first step in my attempt to reorganize the components of TinyStr for better macro ergonomics.

I moved the bytes-to-unsigned functions into a child crate named "tinystr-raw".  This crate does one thing and one thing only: converting from `&[u8]` to one of the three unsigned integer types.

The next step will be to rewire "tinystr-macros" to depend on "tinystr-raw", and finally to make a `macro_rules!` in "tinystr" that ties the pieces together.

This is a 3-crate solution.  I think it is better than the 4-crate solution you had previously tried, because I don't have to re-export any of the types.  Each crate stands on its own with its own specific functions.