@sffc 2020-10-09
Oh yeah, the benchmarks.  They are mixed; some improved, some regressed.  None of the changes are earth-shattering.  Many of the changes are in parts of the code that I didn't touch at all, like `is_ascii_alphanumeric` and `eq`.  Given that the code should be identical, I don't expect that there are any actual problems, but let me know if you think otherwise.

I did attempt to make one change: a separate Error class for the child crate than the one for the parent crate.  This resulted in a measurable regression of 5-20%, which went away when I re-export the child Error class from the parent crate.

Old:

```
     Running target/release/deps/construct-9b243c78efc6f9b5
construct_from_str/4/String                                                                            
                        time:   [374.10 ns 375.93 ns 378.35 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
construct_from_str/4/TinyStr4                                                                            
                        time:   [59.779 ns 60.164 ns 60.623 ns]
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe
construct_from_str/4/TinyStr8                                                                            
                        time:   [107.32 ns 107.50 ns 107.71 ns]
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
construct_from_str/4/TinyStr16                                                                            
                        time:   [132.02 ns 132.45 ns 132.89 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
construct_from_str/4/TinyStrAuto                                                                            
                        time:   [257.36 ns 258.33 ns 259.58 ns]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

construct_from_str/8/String                                                                            
                        time:   [383.63 ns 385.20 ns 387.00 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
construct_from_str/8/TinyStr8                                                                            
                        time:   [109.01 ns 109.39 ns 109.77 ns]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
construct_from_str/8/TinyStr16                                                                            
                        time:   [131.61 ns 132.10 ns 132.62 ns]
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
construct_from_str/8/TinyStrAuto                                                                            
                        time:   [248.26 ns 248.85 ns 249.53 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild

construct_from_str/16/String                                                                            
                        time:   [336.59 ns 339.45 ns 342.94 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe
construct_from_str/16/TinyStr16                                                                            
                        time:   [136.59 ns 137.06 ns 137.66 ns]
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
construct_from_str/16/TinyStrAuto                                                                            
                        time:   [264.53 ns 267.09 ns 270.52 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

construct_from_bytes/4/TinyStr4                                                                            
                        time:   [51.950 ns 52.700 ns 53.687 ns]
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
construct_from_bytes/4/TinyStr8                                                                            
                        time:   [110.16 ns 110.70 ns 111.41 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
construct_from_bytes/4/TinyStr16                                                                            
                        time:   [142.86 ns 143.96 ns 145.28 ns]
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

construct_from_bytes/8/TinyStr8                                                                            
                        time:   [108.92 ns 109.43 ns 109.95 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild
construct_from_bytes/8/TinyStr16                                                                            
                        time:   [139.13 ns 139.39 ns 139.67 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe

construct_from_bytes/16/TinyStr16                                                                            
                        time:   [125.61 ns 126.40 ns 127.54 ns]
Found 12 outliers among 100 measurements (12.00%)
  6 (6.00%) high mild
  6 (6.00%) high severe

construct_unchecked/4/TinyStr4                                                                             
                        time:   [9.4678 ns 9.5270 ns 9.5949 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

construct_unchecked/8/TinyStr8                                                                             
                        time:   [6.8065 ns 6.8290 ns 6.8554 ns]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

construct_unchecked/16/TinyStr16                                                                             
                        time:   [12.486 ns 12.548 ns 12.632 ns]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

     Running target/release/deps/tinystr-a74cb9a3884b8814
convert_to_ascii_lowercase/4/String                                                                            
                        time:   [369.70 ns 371.77 ns 374.22 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
convert_to_ascii_lowercase/4/TinyStr4                                                                             
                        time:   [32.043 ns 32.101 ns 32.161 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
convert_to_ascii_lowercase/4/TinyStr8                                                                             
                        time:   [32.270 ns 32.412 ns 32.569 ns]
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
convert_to_ascii_lowercase/4/TinyStr16                                                                             
                        time:   [42.070 ns 42.177 ns 42.288 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

convert_to_ascii_lowercase/8/String                                                                            
                        time:   [366.04 ns 367.11 ns 368.31 ns]
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low mild
  6 (6.00%) high mild
  2 (2.00%) high severe
convert_to_ascii_lowercase/8/TinyStr8                                                                             
                        time:   [32.422 ns 32.494 ns 32.570 ns]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
convert_to_ascii_lowercase/8/TinyStr16                                                                             
                        time:   [42.470 ns 42.580 ns 42.694 ns]
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  3 (3.00%) high mild
  5 (5.00%) high severe

convert_to_ascii_lowercase/16/String                                                                            
                        time:   [381.99 ns 383.27 ns 384.72 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
convert_to_ascii_lowercase/16/TinyStr16                                                                             
                        time:   [45.504 ns 45.608 ns 45.709 ns]
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe

convert_to_ascii_uppercase/4/String                                                                            
                        time:   [386.20 ns 387.51 ns 388.94 ns]
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) low mild
  4 (4.00%) high mild
convert_to_ascii_uppercase/4/TinyStr4                                                                             
                        time:   [34.424 ns 34.502 ns 34.586 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
convert_to_ascii_uppercase/4/TinyStr8                                                                             
                        time:   [31.896 ns 31.981 ns 32.073 ns]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
convert_to_ascii_uppercase/4/TinyStr16                                                                             
                        time:   [46.967 ns 47.078 ns 47.194 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild

convert_to_ascii_uppercase/8/String                                                                            
                        time:   [424.71 ns 426.25 ns 427.93 ns]
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
convert_to_ascii_uppercase/8/TinyStr8                                                                             
                        time:   [34.201 ns 34.298 ns 34.402 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
convert_to_ascii_uppercase/8/TinyStr16                                                                             
                        time:   [49.546 ns 49.667 ns 49.806 ns]
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  6 (6.00%) high mild
  2 (2.00%) high severe

convert_to_ascii_uppercase/16/String                                                                            
                        time:   [418.43 ns 420.65 ns 423.65 ns]
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
convert_to_ascii_uppercase/16/TinyStr16                                                                             
                        time:   [42.452 ns 42.633 ns 42.840 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low severe
  4 (4.00%) high mild

convert_to_ascii_titlecase/4/String                                                                            
                        time:   [483.08 ns 484.82 ns 486.83 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
convert_to_ascii_titlecase/4/TinyStr4                                                                             
                        time:   [34.397 ns 34.465 ns 34.534 ns]
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe
convert_to_ascii_titlecase/4/TinyStr8                                                                             
                        time:   [36.367 ns 36.505 ns 36.677 ns]
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) low mild
  2 (2.00%) high mild
  3 (3.00%) high severe
convert_to_ascii_titlecase/4/TinyStr16                                                                            
                        time:   [51.289 ns 51.431 ns 51.575 ns]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe

convert_to_ascii_titlecase/8/String                                                                              
                        time:   [485.01 ns 487.46 ns 491.45 ns]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
convert_to_ascii_titlecase/8/TinyStr8                                                                             
                        time:   [36.638 ns 36.737 ns 36.854 ns]
Found 16 outliers among 100 measurements (16.00%)
  8 (8.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe
convert_to_ascii_titlecase/8/TinyStr16                                                                            
                        time:   [51.844 ns 51.939 ns 52.045 ns]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

convert_to_ascii_titlecase/16/String                                                                             
                        time:   [537.36 ns 553.58 ns 570.61 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
convert_to_ascii_titlecase/16/TinyStr16                                                                            
                        time:   [53.161 ns 53.331 ns 53.521 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

test_is_ascii_alphanumeric/4/String                                                                            
                        time:   [73.722 ns 74.376 ns 75.141 ns]
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
test_is_ascii_alphanumeric/4/TinyStr4                                                                             
                        time:   [33.633 ns 34.044 ns 34.499 ns]
Found 24 outliers among 100 measurements (24.00%)
  8 (8.00%) low severe
  10 (10.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
test_is_ascii_alphanumeric/4/TinyStr8                                                                             
                        time:   [46.573 ns 46.682 ns 46.789 ns]
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  5 (5.00%) high severe
test_is_ascii_alphanumeric/4/TinyStr16                                                                            
                        time:   [74.920 ns 75.489 ns 76.257 ns]
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) low mild
  3 (3.00%) high mild
  5 (5.00%) high severe

test_is_ascii_alphanumeric/8/String                                                                            
                        time:   [91.820 ns 92.076 ns 92.335 ns]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
test_is_ascii_alphanumeric/8/TinyStr8                                                                            
                        time:   [49.888 ns 50.028 ns 50.186 ns]
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
test_is_ascii_alphanumeric/8/TinyStr16                                                                            
                        time:   [78.960 ns 79.148 ns 79.337 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild

test_is_ascii_alphanumeric/16/String                                                                            
                        time:   [144.98 ns 145.85 ns 146.76 ns]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe
test_is_ascii_alphanumeric/16/TinyStr16                                                                            
                        time:   [74.762 ns 75.248 ns 75.840 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

test_eq/4/String        time:   [730.19 ns 736.30 ns 745.40 ns]                              
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
test_eq/4/TinyStr4      time:   [120.47 ns 121.01 ns 121.57 ns]                               
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
test_eq/4/TinyStr8      time:   [122.11 ns 122.56 ns 123.28 ns]                               
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe
test_eq/4/TinyStr16     time:   [192.48 ns 192.98 ns 193.51 ns]                                

test_eq/8/String        time:   [377.14 ns 381.39 ns 386.56 ns]                             
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) high mild
  7 (7.00%) high severe
test_eq/8/TinyStr8      time:   [132.44 ns 134.31 ns 136.61 ns]                               
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
test_eq/8/TinyStr16     time:   [216.67 ns 217.57 ns 218.58 ns]                                
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

test_eq/16/String       time:   [306.66 ns 313.15 ns 319.42 ns]                              
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
test_eq/16/TinyStr16    time:   [218.87 ns 221.78 ns 224.53 ns]                                 
```

New:

```
     Running target/release/deps/construct-cbb9633ae911ab5a
construct_from_str/4/String                                                                            
                        time:   [330.58 ns 331.62 ns 332.71 ns]
                        change: [-2.5478% -1.9913% -1.4070%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
construct_from_str/4/TinyStr4                                                                             
                        time:   [40.195 ns 40.425 ns 40.672 ns]
                        change: [+6.5028% +7.7888% +8.7770%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
construct_from_str/4/TinyStr8                                                                            
                        time:   [108.07 ns 108.39 ns 108.71 ns]
                        change: [-1.5091% -0.9326% -0.3698%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
construct_from_str/4/TinyStr16                                                                            
                        time:   [215.75 ns 216.37 ns 217.06 ns]
                        change: [+50.368% +51.133% +52.034%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
…